pyproject.toml
[tool.poetry]
name = "smart-qa"
version = "0.1.0"
description = "Intelligent Q&A tool using Google Gemini AI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "smart_qa"}]

[tool.poetry.dependencies]
python = "^3.9"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
